chaange all f64 to f32 because of opengl

+++ src/math/fct.rs
@@ -1,16 +1,16 @@
-pub fn fast_floor(x : f64) -> i32 {
+pub fn fast_floor(x : f32) -> i32 {
     let xi = x as i32;
-    if x < xi as f64 {
+    if x < xi as f32 {
         return xi - 1;
     }
     return xi;
 }
 
-pub fn soft_min(a : f64, b : f64, k : f64) -> f64 {
+pub fn soft_min(a : f32, b : f32, k : f32) -> f32 {
     let h = (k - (a-b).abs()).max(0.0)/k;
     a.min(b) - h*h*h*k/6.0
 }
 
-pub fn map(val:f64, ai:f64, bi:f64, af:f64, bf:f64) -> f64 {
+pub fn map(val:f32, ai:f32, bi:f32, af:f32, bf:f32) -> f32 {
     af + (val - ai) / (bi-ai) * (bf - af)
 }
