+++ src/math/vec4d.rs
@@ -1,13 +1,13 @@
 #[derive(Copy, Clone, Debug, PartialEq)]
 pub struct Vec4D {
-    pub x : f64,
-    pub y : f64,
-    pub z : f64,
-    pub w : f64,
+    pub x : f32,
+    pub y : f32,
+    pub z : f32,
+    pub w : f32,
 }
 
 impl Vec4D {
-    pub fn new(x: f64, y: f64, z: f64, w: f64) -> Self {
+    pub fn new(x: f32, y: f32, z: f32, w: f32) -> Self {
         Self {
             x : x,
             y : y,
@@ -16,16 +16,16 @@ impl Vec4D {
         }
     }
     
-    pub fn sq_norm(&self) -> f64 {
+    pub fn sq_norm(&self) -> f32 {
         self.x*self.x + self.y*self.y + self.z*self.z + self.w*self.w
     }
 
-    pub fn norm(&self) -> f64 {
+    pub fn norm(&self) -> f32 {
         self.sq_norm().sqrt()
     }
 
     pub fn normalize(&mut self) {
-        let k = 1.0_f64 / self.norm();
+        let k = 1.0_f32 / self.norm();
         self.x *= k;
         self.y *= k;
         self.z *= k;
@@ -33,7 +33,7 @@ impl Vec4D {
     }
 
     pub fn normalized(&self) -> Self {
-        let k = 1.0_f64 / self.norm();
+        let k = 1.0_f32 / self.norm();
         Self {
             x:self.x * k,
             y:self.y * k,
@@ -42,26 +42,26 @@ impl Vec4D {
         }
     }
 
-    pub fn dot(&self, v2 : &Self) -> f64 {
+    pub fn dot(&self, v2 : &Self) -> f32 {
         self.x*v2.x + self.y*v2.y + self.z*v2.z + self.w*v2.w
     }
 
     pub fn null() -> Self {
         Vec4D {
-            x: 0_f64,
-            y: 0_f64,
-            z: 0_f64,
-            w: 0_f64,
+            x: 0_f32,
+            y: 0_f32,
+            z: 0_f32,
+            w: 0_f32,
         }
     }
 
-    pub fn to_vec(&self) -> Vec<f64> {
+    pub fn to_vec(&self) -> Vec<f32> {
         vec![self.x, self.y, self.z, self.w]
     }
 }
 
-impl From<(f64, f64, f64, f64)> for Vec4D {
-    fn from(other: (f64, f64, f64, f64)) -> Self {
+impl From<(f32, f32, f32, f32)> for Vec4D {
+    fn from(other: (f32, f32, f32, f32)) -> Self {
         Vec4D::new(other.0, other.1, other.2, other.3)
     }
 }
@@ -109,9 +109,9 @@ impl std::ops::Neg for Vec4D {
         }
     }
 }
-impl std::ops::Mul<f64> for Vec4D {
+impl std::ops::Mul<f32> for Vec4D {
     type Output = Vec4D;
-    fn mul(self, other : f64) -> Self {
+    fn mul(self, other : f32) -> Self {
         Self{
             x : self.x * other,
             y : self.y * other,
@@ -120,8 +120,8 @@ impl std::ops::Mul<f64> for Vec4D {
         }
     }
 }
-impl std::ops::MulAssign<f64> for Vec4D {
-    fn mul_assign(&mut self, other : f64) {
+impl std::ops::MulAssign<f32> for Vec4D {
+    fn mul_assign(&mut self, other : f32) {
         *self = Self{
             x : self.x * other,
             y : self.y * other,
@@ -130,7 +130,7 @@ impl std::ops::MulAssign<f64> for Vec4D {
         };
     }
 }
-impl std::ops::Mul<Vec4D> for f64 {
+impl std::ops::Mul<Vec4D> for f32 {
     type Output = Vec4D;
     fn mul(self, other : Vec4D) -> Vec4D {
         Vec4D {
